update view private key modal
