+++ src/app/components/modals/private-address-modal/private-address-modal.component.html
@@ -3,15 +3,25 @@
     <h4 class="modal-title">View Secret Key</h4>
   </div>
   <div class="modal-body mt-3">
-    <p>
+    <p *ngIf="!isEtherLinkAddress">
       Public Address
       <a (click)="copyAddress(publicAddress, copyPublAddrButtonRef)" class="copy-address" #copyPublAddrButtonRef data-toggle="tooltip" data-placement="top" title="Click to copy public address"><i class="far fa-copy"></i></a>
       <span class="address">{{publicAddress}}</span>
     </p>
+    <p *ngIf="isEtherLinkAddress">
+      Public EtherLink Address
+      <a (click)="copyAddress(convertToEtherLinkAddress(publicAddress), copyPublEtherLinkAddrButtonRef)" class="copy-address" #copyPublEtherLinkAddrButtonRef data-toggle="tooltip" data-placement="top" title="Click to copy EtherLink address"><i class="far fa-copy"></i></a>
+      <span class="address">{{convertToEtherLinkAddress(publicAddress)}}</span>
+    </p>
+    <p *ngIf="isEtherLinkAddress">
+      Public Ethereum Address
+      <a (click)="copyAddress(convertToEthereumAddress(publicAddress), copyPublEthereumAddrButtonRef)" class="copy-address" #copyPublEthereumAddrButtonRef data-toggle="tooltip" data-placement="top" title="Click to copy Ethereum address"><i class="far fa-copy"></i></a>
+      <span class="address">{{convertToEthereumAddress(publicAddress)}}</span>
+    </p>
     <p>
       Secret Key
-      <a (click)="copyAddress(privateAddress, copyPrivAddrButtonRef)" class="copy-address" #copyPrivAddrButtonRef data-toggle="tooltip" data-placement="top" title="Click to copy public address"><i class="far fa-copy"></i></a>
-      <span class="address">{{privateAddress}}</span>
+      <a (click)="copyAddress(privateAddress, copyPrivAddrButtonRef)" class="copy-address" #copyPrivAddrButtonRef data-toggle="tooltip" data-placement="top" title="Click to copy secret address"><i class="far fa-copy"></i></a>
+      <span class="address">{{privateAddress.slice(0, 32) + '...' + privateAddress.slice(-16)}}</span>
     </p>
   </div>
   <div class="modal-footer justify-content-center mt-2">
